+++ extensions/rest-automation-lib/src/test/resources/public/sample.js
@@ -1,0 +1,3 @@
+var x = 10;
+console.log(x);
+console.log("hello world");