+++ CHANGELOG.md
@@ -6,11 +6,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ---
-## Version 2.2.2, 10/21/2021
+## Version 2.2.2, 10/25/2021
 
 ### Added
 
-N/A
+User defined API authentication functions can be selected using custom HTTP request header
 
 ### Removed
 
@@ -18,7 +18,8 @@ N/A
 
 ### Changed
 
-Improved and streamlined SimpleXmlParser to handle arrays.
+1. Improved and streamlined SimpleXmlParser to handle arrays
+2. Bugfix for file upload in Service Gateway (REST automation library)
 
 ---
 ## Version 2.2.1, 10/1/2021